requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
pymysql==0.9.3
pytest==5.4.2
boto3>=1.17.66
requests>=2.24.0
botocore>=1.20.66
pandas>=1.1.2
scipy>=1.5.2
scikit-learn>=0.23.2
lightfm>=1.16
numpy>=1.19.1